refactor: toPoints 메서드 Points 클래스로 이동
